requirements.txt
fastapi[standard]
uvicorn
pydantic[email]
pymongo[srv]
python-dotenv
pymongo
passlib[bcrypt]
pyjwt
PyGithub
httpx
gitpython
langchain-groq
langchain
cachetools
tiktoken
google-generativeai
certifi
fastapi-limiter